ai/generative-ai-service/decode-images-and-videos-with-oci-gen-ai/files/requirements.txt
streamlit==1.37.0
oci==3.50.1
Pillow
opencv-python-headless==4.10.0.84
tqdm==4.66.3